+++ app/src/modules/settings/routes/extensions/constants/icons.ts
